+++ history.rst
@@ -4,6 +4,17 @@ Simple Python CMS
 
 
 =============
+Version 0.1.2
+=============
+
+2013-02-22
+
+- Weitere Anpassungen von *lib.datadir*
+
+- Erstellen eines Unterordners im Datenordner und zugehörige Tests
+
+
+=============
 Version 0.1.1
 =============
 
@@ -11,7 +22,7 @@ Version 0.1.1
 
 - Die Grundüberlegungen zur Struktur sind abgeschlossen.
 
-- Mit der Datenordner-Verwaltung begonnen. Die Meisten Änderungen befinden sich
+- Mit der Datenordner-Verwaltung begonnen. Die meisten Änderungen befinden sich
   in *lib.datadir*. Über *lib.datadir.tree* kann auf den kompletten
   DATATREEDIR zugegriffen werden.